extract aave write helpers and quote assembly
